transfer across users, log out, validates change username

+++ Interfaces/TransactionInterface.java
@@ -3,7 +3,7 @@ package Interfaces;
 import Types.CurrencyType;
 
 public interface TransactionInterface {
-    public boolean withdraw(float funds, String name);
+    public boolean withdraw(double funds, String name);
 
-    public void deposit(float funds, String name, CurrencyType cType);
+    public void deposit(double funds, String name, CurrencyType cType);
 }
