+++ Objects/Transaction.java
@@ -6,22 +6,22 @@ public class Transaction {
     private TransactionType type;
     private String name;
     private String date;
-    private float amount;
+    private double amount;
 
-    public Transaction(String name, float amount, TransactionType type) {
+    public Transaction(String name, double amount, TransactionType type) {
         this.name = name;
         this.amount = amount;
         this.type = type;
     }
 
-    public Transaction(String name, String date, float amount, TransactionType type) {
+    public Transaction(String name, String date, double amount, TransactionType type) {
         this(name, amount, type);
         this.date = date;
     }
 
     public String getName() { return name; }
 
-    public float getAmount() { return amount; }
+    public double getAmount() { return amount; }
 
     public String getDate() { return date; }
